+++ EventMasters/src/main/java/pl/otez/ticket/dto/TicketRequestDto.java
@@ -2,7 +2,9 @@ package pl.otez.ticket.dto;
 
 import lombok.Builder;
 
-public record TicketRequestDto(Long id, Long client_id, Long price_per_unit, Long event_id) {
+import java.math.BigDecimal;
+
+public record TicketRequestDto(Long id, Long client_id, BigDecimal price_per_unit, Long event_id) {
     @Builder
     public TicketRequestDto{}
 }
